+++ packages/cli/src/controller/build-controller.ts
@@ -2,10 +2,11 @@
 // SPDX-License-Identifier: GPL-3.0
 
 import assert from 'assert';
-import {readFileSync} from 'fs';
+import {readFileSync, existsSync} from 'fs';
 import path from 'path';
 import {globSync} from 'glob';
 import TerserPlugin from 'terser-webpack-plugin';
+import {TsconfigPathsPlugin} from 'tsconfig-paths-webpack-plugin';
 import webpack, {Configuration} from 'webpack';
 import {merge} from 'webpack-merge';
 
@@ -54,6 +55,7 @@ const getBaseConfig = (
 
   resolve: {
     extensions: ['.tsx', '.ts', '.js', '.json'],
+    plugins: [],
   },
 
   output: {
@@ -75,6 +77,14 @@ export async function runWebpack(
     {output: {clean}}
     // Can allow projects to override webpack config here
   );
+  const tsconfigPath = path.join(projectDir, 'tsconfig.json');
+  if (existsSync(tsconfigPath)) {
+    const tsconfig = readFileSync(tsconfigPath, 'utf-8');
+    const tsconfigJson = JSON.parse(tsconfig);
+    if (tsconfigJson.compilerOptions?.paths && config.resolve && config.resolve.plugins) {
+      config.resolve.plugins.push(new TsconfigPathsPlugin());
+    }
+  }
 
   await new Promise((resolve, reject) => {
     webpack(config).run((error, stats) => {