+++ packages/node-core/CHANGELOG.md
@@ -6,6 +6,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [17.2.1] - 2025-03-24
+### Fixed
+- Max queue size error with workers (#2725)
+
+## [17.2.0] - 2025-03-20
+### Fixed
+- Testing not updating test results correctly if fetching block fails (#2716)
+- The `dynamicDatasources` will be cleared in advance, causing subsequent index data to be missing.
+- UnfinalizedBlock rollback not being handled correctly
+
+### Changed
+- On block fetch failure, finish indxing already fetched blocks before exiting (#2720)
+
 ## [17.1.0] - 2025-03-05
 ### Added
 - Flushing the cache every block when the project is fully synced to reduce delay between data indexed and being queryable (#2707)
@@ -936,7 +949,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Move blockchain agnostic code from `node` to `node-core` package. (#1222)
 
-[Unreleased]: https://github.com/subquery/subql/compare/node-core/17.1.0...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node-core/17.2.1...HEAD
+[17.2.1]: https://github.com/subquery/subql/compare/node-core/17.2.0...node-core/17.2.1
+[17.2.0]: https://github.com/subquery/subql/compare/node-core/17.1.0...node-core/17.2.0
 [17.1.0]: https://github.com/subquery/subql/compare/node-core/17.0.2...node-core/17.1.0
 [17.0.2]: https://github.com/subquery/subql/compare/node-core/17.0.1...node-core/17.0.2
 [17.0.1]: https://github.com/subquery/subql/compare/node-core/17.0.0...node-core/17.0.1