+++ packages/node/CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.11.1] - 2025-03-24
+### Fixed
+- Max queue size error with workers (#2725)
+
+## [5.11.0] - 2025-03-20
+### Fixed
+- Testing not updating test results correctly if fetching block fails (#2716)
+
 ## [5.10.0] - 2025-03-05
 ### Changed
 - Update `@subql/node-core` with new cache flush behaviour when fully synced (#2707)
@@ -1394,7 +1402,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - bump @polkadot/api to 3.1.1
 
-[Unreleased]: https://github.com/subquery/subql/compare/node/5.10.0...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/node/5.11.1...HEAD
+[5.11.1]: https://github.com/subquery/subql/compare/node/5.11.0...node/5.11.1
+[5.11.0]: https://github.com/subquery/subql/compare/node/5.10.0...node/5.11.0
 [5.10.0]: https://github.com/subquery/subql/compare/node/5.9.1...node/5.10.0
 [5.9.1]: https://github.com/subquery/subql/compare/node/5.9.0...node/5.9.1
 [5.9.0]: https://github.com/subquery/subql/compare/node/5.8.0...node/5.9.0